application/requirements.txt
streamlit~=1.38.0
streamlit-authenticator~=0.3.3
boto3~=1.33.4
psycopg2-binary==2.9.9
SQLAlchemy==1.4.52
opensearch-py==2.4.2
PyMySQL==1.1.1
python-dotenv~=1.0.0
plotly~=5.18.0
cryptography==42.0.4
langchain~=0.1.11
langchain-core~=0.1.30
sqlparse~=0.4.2
debugpy
pandas==2.0.3
openpyxl
starrocks==1.0.6
clickhouse-sqlalchemy==0.2.6
sagemaker
fastapi~=0.110.1
sqlalchemy-redshift~=0.8.14
numpy==1.26.4
pyhive==0.7.0
thrift==0.20.0
thrift-sasl==0.4.3
sqlalchemy-bigquery==1.11.0
pymssql==2.3.1
PyAthena==3.0.10
pyodps~=0.12.0